backend/app/services/rag_engine.py
import os
from langchain_google_genai import GoogleGenerativeAIEmbeddings, ChatGoogleGenerativeAI
from langchain_community.vectorstores import Chroma
from langchain.chains import RetrievalQA
from dotenv import load_dotenv

load_dotenv()

class SilicaBrain:
    def __init__(self):
        self.embeddings = GoogleGenerativeAIEmbeddings(model="models/embedding-001")
        self.llm = ChatGoogleGenerativeAI(model="gemini-1.5-flash")
        
        # Load existing vector database
        self.vector_db = Chroma(
            persist_directory="app/data/vector_db", 
            embedding_function=self.embeddings
        )

    def ask_tutor(self, question: str):
        qa_chain = RetrievalQA.from_chain_type(
            llm=self.llm,
            chain_type="stuff",
            retriever=self.vector_db.as_retriever()
        )
        
        response = qa_chain.invoke(question)
        return response["result"]